Added image object detection methods

+++ object_detection/utils/predict.py
@@ -9,7 +9,6 @@ import sys
 import subprocess
 import tarfile
 import tensorflow as tf
-import zipfile
 import re
 import json
 import pytesseract
@@ -18,16 +17,16 @@ import os
 from distutils.version import StrictVersion
 from collections import defaultdict
 from io import StringIO
-from PIL import Image, ImageDraw, ImageFont
+from PIL import Image
 from matplotlib import pyplot as plt
-from PIL import Image, ImageFont, ImageDraw, ImageEnhance
-import csv
 import cv2
 import argparse
+import requests
 
 # This is needed since the notebook is stored in the object_detection folder.
 sys.path.append("..")
 sys.path.append("/home/vasanth/mystique/models/research")
+#sys.path.append("/home/keerthanamanoharan/Documents/office_work/Pic2Code/mystique/models/research")
 from object_detection.utils import ops as utils_ops
 
 if StrictVersion(tf.__version__) < StrictVersion('1.9.0'):
@@ -38,8 +37,10 @@ from object_detection.utils import label_map_util
 from object_detection.utils import visualization_utils as vis_util
 
 MODEL_NAME = '/home/vasanth/mystique/object_detection/inference_graph'
+#MODEL_NAME = '/home/keerthanamanoharan/Documents/office_work/Pic2Code/mystique/object_detection/inference_graph'
 PATH_TO_FROZEN_GRAPH = MODEL_NAME + '/frozen_inference_graph.pb'
 PATH_TO_LABELS = '/home/vasanth/mystique/object_detection/training/object-detection.pbtxt'
+#PATH_TO_LABELS = '/home/keerthanamanoharan/Documents/office_work/Pic2Code/gitlab/mystique/object_detection/training/object-detection.pbtxt'
 
 
 detection_graph = tf.Graph()
@@ -52,7 +53,72 @@ with detection_graph.as_default():
     
 category_index = label_map_util.create_category_index_from_labelmap(PATH_TO_LABELS, use_display_name=True)
 
-from PIL import Image, ImageDraw
+
+"""[checks for overlapping rectangels in the countours]
+
+Returns:
+    [Boolean] -- [if overlaps or not]
+"""
+def contains(point, coords):
+  return (float(coords[0])<=point[0]+3<=float(coords[2])) and (float(coords[1])<=point[1]+3<=float(coords[3]))
+
+"""[image hosting service]
+
+Returns:
+    [list] -- [image urls , coords]
+"""
+def image_crop_get_url(coords_list,img_pillow):
+  images=[] 
+  image_coords=[] 
+  for coords in coords_list:
+    cropped=img_pillow.crop((coords[0],coords[1],coords[0]+coords[2],coords[1]+coords[3]))
+    cropped.save("tmep_image.png")
+    img=open("tmep_image.png","rb").read()
+    base64_string=base64.b64encode(img).decode()
+    #reponse=requests.post(url="https://api.imgur.com/3/image",data={"image":base64_string},headers={"Authorization":"Client-ID 017c383a18f0d47"})
+    #images.append(reponse.json().get('data',{}).get('link',''))
+    images.append("")
+    image_coords.append(coords)
+  return images,coords_list
+
+"""[Returns the dected images coords and hosted urls]
+
+Returns:
+    [list] -- [results from image_crop_list]
+"""
+def image_detection(img,detected_coords,img_pillow):
+    points=[]
+    #pre processing
+    gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
+    dst = cv2.equalizeHist(gray)
+    blur = cv2.GaussianBlur(dst,(5,5),0)
+    #edge detection
+    edged = cv2.Canny(blur,0,255)
+    #countours
+    _, contours, hierarchy = cv2.findContours(edged, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
+    #get the coords of the contours
+    for c in contours:
+        (x,y,w,h) = cv2.boundingRect(c)
+        points.append((x,y,w,h))
+    
+    #extarct the points that lies inside the detected objects coords [ rectangle ]
+    included_points_positions=[0] * len(points)
+    for point in points:
+      for p in detected_coords:
+        if contains((point[0],point[1]),p):
+          included_points_positions[points.index(point)]=1
+    image_points=[]
+    # now get the image points / coords that lies outside the detected objects coords     
+    for point in points:
+      if included_points_positions[points.index(point)]!=1:
+        image_points.append(point)
+    image_points=sorted(set(image_points),key=image_points.index)
+    image_points=image_points[:-1]
+
+    return image_crop_get_url(image_points,img_pillow)
+    
+
+
 
 def run_inference_for_single_image(image, graph,tensor_dict,sess):
     
@@ -87,9 +153,66 @@ def run_inference_for_single_image(image, graph,tensor_dict,sess):
         output_dict['detection_masks'] = output_dict['detection_masks'][0]
     return output_dict
 
-def get_card_json(objects):
-        body=[]
+"""[returns the grouped image objects as adaptive card json body]
+
+Returns:
+    [None] -- [apends ymins and body]
+"""
+def get_image_objects(image_objects,body,ymins):
+
+    #group the image objects based on ymin
+    groups=[]
+    unique_ymin=list(set([x.get('ymin') for x in image_objects]))
+    for un in unique_ymin:
+        l=[]
+        for xx in image_objects:
+            if xx.get('ymin')==un:
+                l.append(xx)
+        groups.append(l)
+    
+    #now put similar ymin grouped objects into a imageset - if a group has more than one image object
+    for group in groups:
+        if len(group)==1:
+            object=group[0]
+            body.append( {
+                "type": "Image",
+                "altText": "Image",
+                "horizontalAlignment": object.get('horizontal_alignment',''),
+                "url":object.get('url'),
+                "coords":object.get('coords','')
+                })
+            ymins.append(object.get('ymin'))
+        else:
+            image_set= {
+            "type": "ImageSet",
+            "imageSize": "medium",
+            "images": []}
+            for object in group:
+                obj= {"type": "Image",
+                "altText": "Image",
+                "horizontalAlignment": object.get('horizontal_alignment',''),
+                "url":object.get('url'),
+                "coords":object.get('coords','')
+                }
+                image_set['images'].append(obj)
+
+            body.append(image_set)
+            ymins.append(object.get('ymin'))
+
+
+"""[Returns the adapative card json body based on the detection]
+
+Returns:
+    [list] -- [body, ymins]
+"""
+def get_card_json(objects,images_number):
+        body=[] 
+        ymins=[]
+        image_objects=[]           
         for object in objects:
+            if object.get('object')=="image":
+                image_objects.append(object)
+    
             if object.get('object')=="textbox":
 
                 if len(object.get('text','').split())>=10:
@@ -107,6 +230,7 @@ def get_card_json(objects):
                         }
                         ]
                         })
+                    ymins.append(object.get('ymin'))
                 else:
                     body.append({
                     "type": "TextBlock",
@@ -117,12 +241,14 @@ def get_card_json(objects):
                     "weight":object.get('weight',''),
                     "coords":object.get('coords','')
                     })
+                    ymins.append(object.get('ymin'))
             if object.get('object')=="checkbox":
                 body.append({
                     "type": "Input.Toggle",
                     "title": object.get('text',''),
                     "coords":object.get('coords','')
                     })
+                ymins.append(object.get('ymin'))
             if object.get('object')=="radio_button":
                 body.append( {
                     "type": "Input.ChoiceSet",
@@ -135,23 +261,30 @@ def get_card_json(objects):
                         ],
                      "style": "expanded"
                     })
-            if object.get('object')=="image":
-                body.append( {
-                    "type": "Image",
-                    "altText": "",
-                    "coords":object.get('coords','')
-                    })
-        return body
+                ymins.append(object.get('ymin'))
+        get_image_objects(image_objects,body,ymins)
+        return body,ymins
 
+"""[OCR to get the text of the detected coords]
+
+Returns:
+    [string] -- [OCR text]
+"""
 def get_text(image, coords):
     cropped_image = image.crop(coords)
     data = pytesseract.image_to_string(cropped_image, lang='eng',config='--psm 6')
     return data
 
+"""[Returns the size and weight properites of an object]
+
+Returns:
+    [string] -- [size,weight]
+"""
 def get_size_and_weight(image,coords):
     cropped_image = image.crop(coords)
     cropped_image.save("temp_image.png")
     img=cv2.imread("temp_image.png")
+    #preprocess
     gray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
     gray = cv2.blur(gray, (5,5))
 
@@ -159,11 +292,14 @@ def get_size_and_weight(image,coords):
     closing = cv2.morphologyEx(gray, cv2.MORPH_CLOSE, kernel)
 
     edged = cv2.Canny(img, 30, 200)
+    #edge detection
     cv2.imshow("Canny", edged)
+    #contours bulding
     _, contours, _ = cv2.findContours(edged, cv2.RETR_LIST, cv2.CHAIN_APPROX_SIMPLE)
     box_width=[]
     box_height=[]
 
+    #calculate the average width and height of the contour coords of the object
     for c in contours:
         rect = cv2.minAreaRect(c)
         (x,y,w,h) = cv2.boundingRect(c)
@@ -198,7 +334,13 @@ def get_size_and_weight(image,coords):
     
     return size,weight
 
+"""[Return the horizontal alignment of the object]
+
+Returns:
+    [string] 
+"""
 def get_alignment(image,xmin,xmax):
+    
     avg=math.ceil((xmin+xmax)/2)
     w,h=image.size
     if (avg/w)*100 >=0 and (avg/w)*100 <45:
@@ -208,17 +350,28 @@ def get_alignment(image,xmin,xmax):
     else:
         return "Right"
 
+"""[Returns the distance between 2 points  [ used for difference in RGB for color detecion ]]
+
+Returns:
+    [ndarray] 
+"""
 def get_distance(v1, v2):
     return np.sqrt(np.sum((v1 - v2) ** 2))
 
+"""[Returns the text color of the object [ mainly textboxes and rich textboxes ]]
+
+Returns:
+    [text] 
+"""
 def get_colors(image,coords):
     cropped_image = image.crop(coords)
     cropped_image.save("temp_image.png")
+    # get 2 dominant colors
     q = cropped_image.quantize(colors=2,method=2)
     dominant_color= q.getpalette()[3:6]
     
     colors={
-        "Attention":(255,0,0),
+        "get_distance":(255,0,0),
         "Accent":(0,0,255),
         "Good":(0,128,0),
         "Dark":(0,0,0),
@@ -228,11 +381,13 @@ def get_colors(image,coords):
     color='Default'
     found_colors=[]
     distances=[]
+    #find the dominant text colors based on the RGB difference
     for key,values in colors.items():
         distance=get_distance(np.asarray(values),np.asarray(dominant_color))
         if distance<=150:
             found_colors.append(key)
             distances.append(distance)
+    #If the color is predicted as LIGHT check for false cases where both dominan colors are White
     if found_colors!=[]:
         index=distances.index(min(distances))
         color=found_colors[index]
@@ -283,7 +438,10 @@ def main(input_file_path):
                     classes=output_dict['detection_classes']                        
                     r,c=boxes.shape
                     json_object={"file_name":image}
+                    detected_coords=[]
                     json_object['objects']=[]
+
+                    #For detected objects
                     for i in range(r):
                       
                       if scores[i]*100>=60.0:
@@ -294,17 +452,18 @@ def main(input_file_path):
                             object_json['object']="radio_button"
                         elif  str(classes[i])=="3":
                             object_json['object']="checkbox"
-                        elif  str(classes[i])=="4":
-                            object_json['object']="image"
+                        
                         ymin = boxes[i][0]*height
                         xmin = boxes[i][1]*width
                         ymax = boxes[i][2]*height
                         xmax = boxes[i][3]*width
+                        
                         object_json['xmin']=str(xmin)
                         object_json['ymin']=str(ymin)
                         object_json['xmax']=str(xmax)
                         object_json['ymax']=str(ymax)
                         object_json['coords']=','.join([str(xmin),str(ymin),str(xmax),str(ymax)])
+                        detected_coords.append((xmin,ymin,xmax,ymax))
                         object_json['text']=get_text(image_pillow,((xmin, ymin, xmax,ymax )))
                         if object_json['object']=="textbox":
                             
@@ -312,8 +471,25 @@ def main(input_file_path):
                             object_json["horizontal_alignment"]=get_alignment(image_pillow,float(xmin),float(xmax))
                             object_json['color']=get_colors(image_pillow,((xmin, ymin, xmax,ymax )))
                         json_object['objects'].append(object_json)
+                    
+                    #For image objects
+                    images,image_coords=image_detection(image_np,detected_coords,image_pillow)
+                    ctr=0
+                    for im in images:
+                        coords=image_coords[ctr]
+                        coords=(coords[0],coords[1],coords[0]+coords[2],coords[1]+coords[3])
+                        object_json=dict().fromkeys(['object','xmin','ymin','xmax','ymax'],'')
+                        object_json["object"]="image"
+                        object_json["horizontal_alignment"]=get_alignment(image_pillow,float(coords[0]),float(coords[2]))
+                        object_json["url"]=im
+                        object_json['xmin']=coords[0]
+                        object_json['ymin']=coords[1]
+                        object_json['xmax']=coords[2]
+                        object_json['ymax']=coords[3]
+                        object_json['coords']=','.join([str(coords[0]),str(coords[1]),str(coords[2]),str(coords[3])])
+                        json_object['objects'].append(object_json)
+                        ctr+=1
 
- 
                     vis_util.visualize_boxes_and_labels_on_image_array(
                         image_np,
                         output_dict['detection_boxes'],
@@ -322,21 +498,31 @@ def main(input_file_path):
                         category_index,
                         instance_masks=output_dict.get('detection_masks'),
                         use_normalized_coordinates=True,
-                        line_thickness=5,
-                        skip_scores=False,
+                        line_thickness=2,
+                        skip_scores=True,
+                        skip_labels=True,
                         min_score_thresh=0.6
                         )
                     return_dict["image"]=base64.b64encode(cv2.imencode('predicted.jpg', image_np)[1]).decode("utf-8")
-                    for obj in range(len(json_object.get('objects'))-1,0,-1):
-                        for i in range(obj):
-                            if float(json_object.get('objects')[i].get('ymin'))>float(json_object.get('objects')[i+1].get('ymin')):
-                                temp = json_object.get('objects')[i]
-                                json_object.get('objects')[i] = json_object.get('objects')[i+1]
-                                json_object.get('objects')[i+1] = temp
-                                
+                   
+                    
                     card_json = {"type": "AdaptiveCard", "version": "1.0", "body": [], "$schema": "http://adaptivecards.io/schemas/adaptive-card.json"}
-                    card_json["body"]=get_card_json(json_object.get('objects',[]))
-                    os.remove('temp_image.png')
+                    body,ymins=get_card_json(json_object.get('objects',[]),len(images))
+                    
+                    #Vertical Alignment of the  card objects [  based on ymin ]
+                    for obj in range(len(ymins)-1,0,-1):
+                        for i in range(obj):
+                            if float(ymins[i])>float(ymins[i+1]):
+                                temp1=ymins[i]
+                                temp = body[i]
+                                body[i] = body[i+1]
+                                ymins[i]=ymins[i+1]
+                                body[i+1] = temp
+                                ymins[i+1] = temp1
+                    card_json["body"]=body
+
+                    if os.path.exists("tmep_image.png"):
+                        os.remove('temp_image.png')
 
                     return_dict["card_json"]=card_json
                     print(json.dumps(return_dict))
